+++ lib/src/core/utils/validator.dart
@@ -31,7 +31,7 @@ String? validatePassword(String? value, String password) {
     return 'Please enter password';
   } else if (value.length < 8) {
     return 'Password must be 8 characters long';
-  }  else {
+  } else {
     return null;
   }
 }