+++ lib/src/features/auth/biometric/bloc/biometric_state.dart
@@ -47,14 +47,15 @@ class BiometricsStatusCheckedSuccessState extends BiometricState {
   BiometricsStatusCheckedSuccessState(this.isEnabled);
 }
 
-class RegisteringBiometricsState  extends BiometricState{
-}
-class RegisterBiometricsFailureState extends BiometricState{
+class RegisteringBiometricsState extends BiometricState {}
+
+class RegisterBiometricsFailureState extends BiometricState {
   final String message;
 
   RegisterBiometricsFailureState(this.message);
 }
-class RegisterBiometricsSuccessState extends BiometricState{
+
+class RegisterBiometricsSuccessState extends BiometricState {
   final String message;
 
   RegisterBiometricsSuccessState(this.message);