+++ lib/src/features/auth/presentation/bloc/login_bloc/login_bloc.dart
@@ -35,8 +35,7 @@ class LoginBloc extends Bloc<LoginEvent, LoginState> {
           yield LoginSuccessState(response);
           return;
         }
-        yield LoginFailureState(
-            errorMessage: response.message ?? 'An unknown error occurred.');
+        yield LoginFailureState(errorMessage: response.message);
       } catch (e) {
         yield LoginFailureState(errorMessage: e.toString());
       }