+++ lib/src/features/auth/social/data/repository/social_auth_repository.dart
@@ -30,24 +30,11 @@ class SocialAuthRepository implements ISocialAuthRepository {
   }
 
   Future<SocialModel> facebookLogin() async {
-    // FacebookLogin facebookSignIn = FacebookLogin();
     final LoginResult? result = await FacebookAuth.instance
         .login(); // by default we request the email and the public profile
-
-    // if (result.status == LoginStatus.success) {
-    //   // facebookSignIn.logOut();
-    //   debugPrint('loggout from facebook');
-    // }
-    // final FacebookLoginResult result = await facebookSignIn.logIn([
-    //   'email',
-    // ]);
-
-    // facebookSignIn.loginBehavior = FacebookLoginBehavior.nativeWithFallback;
     switch (result!.status) {
       case LoginStatus.success:
         final AccessToken? accessToken = result.accessToken;
-
-        // final FacebookAccessToken accessToken = result.accessToken;
         SocialModel model =
             await repository.loginWithFacebook(token: accessToken!.token);
         debugPrint(accessToken.expires.toString());
@@ -69,16 +56,11 @@ class SocialAuthRepository implements ISocialAuthRepository {
           await googleSignIn.signIn();
       GoogleSignInAuthentication authentication =
           await googleSignInAccount!.authentication;
-      debugPrint('AuthCode ${authentication.accessToken}');
       final model =
           await repository.loginWithGoogle(token: authentication.accessToken!);
       return model;
     } catch (e) {
-      // rethrow;
-      print("error===>${e}");
-      // return null;
-      return SocialModel(
-          ok: false, message: 'User has cancel login with google');
+      return SocialModel(ok: false, message: '${e.toString()}');
     }
   }
 }