+++ lib/src/features/user/presentation/bloc/user_bloc.dart
@@ -4,7 +4,6 @@ import 'package:bloc/bloc.dart';
 import 'package:injectable/injectable.dart';
 import 'package:meta/meta.dart';
 import 'package:super_module/src/features/auth/data/models/user_model.dart';
-import 'package:super_module/src/features/user/data/model/user_rank.dart';
 import 'package:super_module/src/features/user/data/model/user_rank_model.dart';
 import 'package:super_module/src/features/user/domain/controller/user_controller.dart';